added bpm sorting function
note: i'm a little confused about how this will work in the end (kinda just figuring it out as i go).

idea(?): object of this class will be created, and once daily, all of the appropriate functions will be called, and the data will be stored in the mongodb database. from there, the website can access the database to display the information without needing to perform time-consuming operations every time the page is rendered.